services/gateway/app/tracing.py
from opentelemetry import trace
from opentelemetry.sdk.resources import SERVICE_NAME, Resource
from opentelemetry.sdk.trace import TracerProvider
from opentelemetry.sdk.trace.export import BatchSpanProcessor
from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter


def setup_tracing(service_name: str):
    provider = TracerProvider(resource=Resource.create({SERVICE_NAME: service_name}))
    trace.set_tracer_provider(provider)

    otlp_exporter = OTLPSpanExporter(
        endpoint="http://jaeger:4318/v1/traces",  # HTTP OTLP endpoint
        insecure=True  # indispensable si HTTPS pas configuré
    )

    provider.add_span_processor(BatchSpanProcessor(otlp_exporter))
