services/requirements.txt
# fastapi requirements
fastapi
uvicorn[standard]
httpx

# kafka
aiokafka

# opentelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-asyncio
opentelemetry-instrumentation-aiokafka
opentelemetry-proto


# prometheus
prometheus-fastapi-instrumentator

